ai_sticker_backend/requirements.txt
fastapi
uvicorn
pydantic
Pillow
python-multipart
torch          # GPU recommended
diffusers[torch]
transformers
rembg
transformers
torchaudio

